Feat: 나의 매칭 조회 API, 자동 매칭 중복 방지 로직 추가 및 매칭 결과 API 수정 #92

+++ src/main/java/com/dnd/Exercise/domain/field/dto/response/FindFieldResultDto.java
@@ -14,7 +14,7 @@ public class FindFieldResultDto {
 
     private String profileImg;
 
-    private double totalScore;
+    private int totalScore;
 
     private Integer totalRecordCount;
 
@@ -34,7 +34,7 @@ public class FindFieldResultDto {
         this.totalExerciseTimeMinute = score.get(3);
     }
 
-    public void addTotalScore(double score){
+    public void addTotalScore(int score){
         this.totalScore += score;
     }
 }
